The big tools refactor.
Now turned into Substance.Component classes.
