+++ surface/index.js
@@ -11,12 +11,6 @@ Surface.NodeView = require('./node_view');
 Surface.AnnotationView = require('./annotation_view');
 Surface.TextProperty = require('./text_property');
 
-Surface.Tool = require('./tool');
-Surface.AnnotationTool = require('./annotation_tool');
-Surface.SwitchTypeTool = require('./switch_type_tool');
-Surface.ToolRegistry = require('./tool_registry');
 Surface.Panel = require('./panel');
-
-Surface.Tools = require('./tools');
 
 module.exports = Surface;
