--- surface/tools/strong_tool.js
@@ -1,9 +1,0 @@
-'use strict';
-
-var AnnotationTool = require('../annotation_tool');
-
-var StrongTool = AnnotationTool.extend({
-  name: "strong"
-});
-
-module.exports = StrongTool;
