+++ ui/tools/emphasis_tool.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var AnnotationTool = require('../annotation_tool');
+var AnnotationTool = require('./annotation_tool');
 
 var EmphasisTool = AnnotationTool.extend({
   name: "emphasis"
