+++ ui/tools/strong_tool.js
@@ -1,0 +1,9 @@
+'use strict';
+
+var AnnotationTool = require('./annotation_tool');
+
+var StrongTool = AnnotationTool.extend({
+  name: "strong"
+});
+
+module.exports = StrongTool;
